# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AdminMovieDetailResponse:
    properties:
      cast_ids:
        items:
          type: integer
        type: array
      dates:
        items:
          type: string
        type: array
      director_ids:
        items:
          type: integer
        type: array
      duration:
        type: string
      genre_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      location_ids:
        items:
          type: integer
        type: array
      poster:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      times:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  dto.AdminMovieListDTO:
    properties:
      duration:
        type: string
      genres:
        type: string
      id:
        type: integer
      poster:
        type: string
      release_date:
        type: string
      title:
        type: string
    type: object
  dto.AdminMovieListResponse:
    properties:
      limit:
        type: integer
      movies:
        items:
          $ref: '#/definitions/dto.AdminMovieListDTO'
        type: array
      page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
  dto.AdminResponseError:
    properties:
      message:
        example: Error message detail
        type: string
      status:
        example: error
        type: string
    type: object
  dto.AdminResponseSuccess:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: limustadji@gmail.com
    name: alpha male
    url: https://github.com/L1mus
  description: This is a backend project for a web application called E-wallet.
  license:
    name: MIT
  title: Movie API
  version: "1.0"
paths:
  /admin/movies:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua film dengan support searching, filtering,
        dan pagination
      parameters:
      - description: 'Halaman aktif (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah data per halaman (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Filter Bulan
        in: query
        name: month
        type: integer
      - description: Filter Tahun
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses mengambil list film
          schema:
            allOf:
            - $ref: '#/definitions/dto.AdminResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminMovieListResponse'
              type: object
        "400":
          description: Query parameter tidak valid
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
      summary: Get Admin List Movies
      tags:
      - Admin
    post:
      consumes:
      - multipart/form-data
      description: Menambahkan data film baru beserta file poster dan otomatisasi
        jadwal tayang
      parameters:
      - description: Judul Film
        in: formData
        name: title
        required: true
        type: string
      - description: File Gambar Poster Film
        in: formData
        name: poster
        required: true
        type: file
      - description: 'Tanggal Rilis (Format: YYYY-MM-DD)'
        in: formData
        name: release_date
        required: true
        type: string
      - description: Durasi Jam
        in: formData
        name: duration_hour
        required: true
        type: integer
      - description: Durasi Menit
        in: formData
        name: duration_minute
        required: true
        type: integer
      - description: Sinopsis
        in: formData
        name: synopsis
        required: true
        type: string
      - collectionFormat: multi
        description: Array ID Genre
        in: formData
        items:
          type: integer
        name: genre_ids
        required: true
        type: array
      - collectionFormat: multi
        description: Array ID Cast
        in: formData
        items:
          type: integer
        name: cast_ids
        required: true
        type: array
      - collectionFormat: multi
        description: Array ID Sutradara
        in: formData
        items:
          type: integer
        name: director_ids
        required: true
        type: array
      - collectionFormat: multi
        description: Array ID Kota/Lokasi Bioskop
        in: formData
        items:
          type: integer
        name: location_ids
        required: true
        type: array
      - collectionFormat: multi
        description: Array Tanggal (YYYY-MM-DD)
        in: formData
        items:
          type: string
        name: dates
        required: true
        type: array
      - collectionFormat: multi
        description: Array Jam Tayang (HH:MM)
        in: formData
        items:
          type: string
        name: times
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Sukses menambahkan film baru
          schema:
            $ref: '#/definitions/dto.AdminResponseSuccess'
        "400":
          description: Payload input atau file tidak valid
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
      summary: Add New Movie
      tags:
      - Admin
  /admin/movies/{id}:
    get:
      consumes:
      - application/json
      description: Mengambil data lengkap satu film berdasarkan ID, termasuk semua
        relasi (genre, cast, director, location, dates, times). Digunakan untuk mengisi
        form edit movie.
      parameters:
      - description: ID Film
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sukses mengambil detail film
          schema:
            allOf:
            - $ref: '#/definitions/dto.AdminResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/dto.AdminMovieDetailResponse'
              type: object
        "400":
          description: ID tidak valid
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "404":
          description: Film tidak ditemukan
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
      summary: Get Movie Detail
      tags:
      - Admin
    patch:
      consumes:
      - multipart/form-data
      description: Mengubah data film. Biarkan input kosong jika tidak ingin mengubah
        data spesifik.
      parameters:
      - description: ID Film yang ingin diedit
        in: path
        name: id
        required: true
        type: integer
      - description: Judul Film Baru
        in: formData
        name: title
        type: string
      - description: File Gambar Poster Baru
        in: formData
        name: poster
        type: file
      - description: Tanggal Rilis Baru (YYYY-MM-DD)
        in: formData
        name: release_date
        type: string
      - description: Durasi Jam Baru
        in: formData
        name: duration_hour
        type: integer
      - description: Durasi Menit Baru
        in: formData
        name: duration_minute
        type: integer
      - description: Sinopsis Baru
        in: formData
        name: synopsis
        type: string
      - collectionFormat: multi
        description: Array ID Genre Baru
        in: formData
        items:
          type: integer
        name: genre_ids
        type: array
      - collectionFormat: multi
        description: Array ID Cast Baru
        in: formData
        items:
          type: integer
        name: cast_ids
        type: array
      - collectionFormat: multi
        description: Array ID Sutradara Baru
        in: formData
        items:
          type: integer
        name: director_ids
        type: array
      - collectionFormat: multi
        description: Array ID Lokasi Baru
        in: formData
        items:
          type: integer
        name: location_ids
        type: array
      - collectionFormat: multi
        description: Array Tanggal Baru (YYYY-MM-DD)
        in: formData
        items:
          type: string
        name: dates
        type: array
      - collectionFormat: multi
        description: Array Waktu Baru (HH:MM)
        in: formData
        items:
          type: string
        name: times
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Sukses mengubah data film
          schema:
            $ref: '#/definitions/dto.AdminResponseSuccess'
        "400":
          description: Payload input tidak valid
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "404":
          description: Film tidak ditemukan
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.AdminResponseError'
      summary: Edit Movie
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token used for authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"